restructure some code
