+++ examples/src/scenes/SimpleScene.h
@@ -9,7 +9,7 @@
 #ifndef SIMPLESCENE_H
 #define SIMPLESCENE_H
 
-#include "../../../src/core/GameObject.h"
+#include "../../../src/core/ecs/GameObject.h"
 #include "../../../src/core/Scene.h"
 
 class SimpleScene : public Scene {
