+++ src/core/Project.cpp
@@ -1,0 +1,9 @@
+/**
+ * @file
+ * @brief
+ * @details
+ * @author  Nur Akmal bin Jalil
+ * @date    2025-05-19
+ */
+
+#include "Project.h"
