+++ src/core/Project.h
@@ -1,0 +1,21 @@
+/**
+ * @file
+ * @brief
+ * @details
+ * @author  Nur Akmal bin Jalil
+ * @date    2025-05-19
+ */
+
+
+#ifndef PROJECT_H
+#define PROJECT_H
+
+
+
+class Project {
+
+};
+
+
+
+#endif //PROJECT_H
