+++ src/core/camera/Camera.h
