+++ src/core/camera/CameraManager.cpp
