+++ src/core/camera/CameraManager.h
