+++ src/core/camera/UICamera.cpp
