+++ src/core/camera/UICamera.h
